: Finalizing the project
I finished the project adding the styles to the html, precisely to the main and footer, that's what I was missing.
I followed good practices, I also added a link to the normalize stylesheet so that it looks good in all browsers.
I updated the readme.md and that should do it.